- move generator to react-lowcode - create  simple entry point for generation
