+++ packages/react-lowcode/src/codegen/public/favicon.ico
