+++ packages/react-lowcode/src/codegen/public/logo512.png
