+++ packages/react-lowcode/src/codegen/public/manifest.json
