+++ packages/react-lowcode/src/codegen/src/definition/detail-definition-core.ts
