+++ packages/react-lowcode/src/codegen/src/definition/grommet/detail.ts
