+++ packages/react-lowcode/src/codegen/src/generation/generators/detail/detail-generator-factory.ts
