+++ packages/react-lowcode/src/codegen/src/generation/generators/detail/grommet-detail-generator.ts
