+++ packages/react-lowcode/src/codegen/src/generation/react-components/react-intl/intl-formatter.ts
