+++ packages/react-lowcode/src/codegen/src/generation/typeAlias.ts
