+++ packages/react-lowcode/src/codegen/src/tests/detail/detail.test.ts
@@ -1,7 +1,7 @@
 // TODO https://github.com/vvakame/typescript-formatter/blob/master/lib/formatter.ts
 import { Project, SourceFile } from "ts-morph"
 import ts, { factory } from "typescript"
-import { graphqlGenTs1 } from "../typeAlias.example"
+import { graphqlGenTs1 } from "../../../typeAlias.example"
 import { UiFramework, TableType } from '../../definition/context-types'
 import { ModuleGenerator } from '../../generation/generators/module-generator'
