+++ packages/react-lowcode/src/codegen/tsconfig.json
@@ -22,5 +22,5 @@
   },
   "include": [
     "src"
-  ]
+, "typeAlias.example.ts"  ]
 }
