+++ packages/react-lowcode/src/codegen/typeAlias.example.ts
